code sanitization (decentralized several vars, 2)
