+++ include/main.h
@@ -21,5 +21,6 @@
 #include "corona.h"
 #include "boot.h"
 #include "libpax_helpers.h"
+#include "power.h"
 
 #endif
