+++ include/sensor.h
@@ -1,6 +1,8 @@
 #ifndef _SENSOR_H
 #define _SENSOR_H
 
+#include "configmanager.h"
+
 #define HAS_SENSORS (HAS_SENSOR_1 || HAS_SENSOR_2 || HAS_SENSOR_3)
 
 uint8_t sensor_mask(uint8_t sensor_no);
